refactor(wasm-dpp): eliminate duplicated code
