+++ packages/wasm-dpp/src/identity/identity_facade.rs
@@ -1,4 +1,3 @@
-use serde_json::Value;
 use std::convert::TryInto;
 
 use std::sync::Arc;
@@ -6,28 +5,26 @@ use std::sync::Arc;
 use wasm_bindgen::prelude::*;
 
 use dpp::identity::validation::PublicKeysValidator;
-use dpp::identity::{Identity, IdentityFacade, IdentityPublicKey, KeyID};
+use dpp::identity::{Identity, IdentityFacade};
 
 use crate::bls_adapter::BlsAdapter;
 use crate::buffer::Buffer;
 use crate::errors::{from_dpp_err, RustConversionError};
 use crate::identifier::IdentifierWrapper;
 use crate::identity::errors::InvalidIdentityError;
-use crate::identity::identity_public_key_transitions::IdentityPublicKeyCreateTransitionWasm;
-use crate::utils::{generic_of_js_val, to_vec_of_serde_values};
+
+use crate::utils::ToSerdeJSONExt;
 use crate::validation::ValidationResultWasm;
 use crate::{
-    create_asset_lock_proof_from_wasm_instance, utils, with_js_error, ChainAssetLockProofWasm,
-    IdentityCreateTransitionWasm, IdentityPublicKeyWasm, IdentityTopUpTransitionWasm,
-    IdentityUpdateTransitionWasm, IdentityWasm, InstantAssetLockProofWasm,
+    create_asset_lock_proof_from_wasm_instance, with_js_error, ChainAssetLockProofWasm,
+    IdentityCreateTransitionWasm, IdentityTopUpTransitionWasm, IdentityUpdateTransitionWasm,
+    IdentityWasm, InstantAssetLockProofWasm,
 };
 use dpp::dashcore::{consensus, InstantLock, Transaction};
 
-use dpp::identity::state_transition::identity_public_key_transitions::IdentityPublicKeyCreateTransition;
 use dpp::version::ProtocolVersionValidator;
 use dpp::NonConsensusError;
 use serde::Deserialize;
-use wasm_bindgen::__rt::Ref;
 
 #[derive(Clone)]
 #[wasm_bindgen(js_name=IdentityFacade)]
@@ -57,15 +54,8 @@ impl IdentityFacadeWasm {
         asset_lock_proof: JsValue,
         public_keys: js_sys::Array,
     ) -> Result<IdentityWasm, JsValue> {
-        let asset_lock_proof = create_asset_lock_proof_from_wasm_instance(&asset_lock_proof)?;
-
-        let raw_public_keys = to_vec_of_serde_values(public_keys.iter())?;
-
-        let public_keys = raw_public_keys
-            .into_iter()
-            .map(|v| IdentityPublicKey::from_raw_object(v).map(|key| (key.id, key)))
-            .collect::<Result<_, _>>()
-            .map_err(|e| format!("converting to collection of IdentityPublicKeys failed: {e:#}"))?;
+        let (asset_lock_proof, public_keys) =
+            super::factory_utils::parse_create_args(asset_lock_proof, public_keys)?;
 
         self.0
             .create(asset_lock_proof, public_keys)
@@ -85,9 +75,7 @@ impl IdentityFacadeWasm {
             Default::default()
         };
 
-        let identity_json = utils::stringify(&identity_object)?;
-        let raw_identity: Value =
-            serde_json::from_str(&identity_json).map_err(|e| e.to_string())?;
+        let raw_identity = identity_object.with_serde_to_json_value()?;
 
         let result = self
             .0
@@ -209,58 +197,8 @@ impl IdentityFacadeWasm {
         identity: &IdentityWasm,
         public_keys: &JsValue,
     ) -> Result<IdentityUpdateTransitionWasm, JsValue> {
-        let mut add_public_keys = None;
-
-        if js_sys::Reflect::has(public_keys, &"add".into()).unwrap_or(false) {
-            let raw_add_public_keys = js_sys::Reflect::get(public_keys, &"add".into()).unwrap();
-
-            let add_public_keys_array: &js_sys::Array = raw_add_public_keys
-                .dyn_ref::<js_sys::Array>()
-                .ok_or_else(|| {
-                    RustConversionError::Error(String::from("public keys to add must be array"))
-                        .to_js_value()
-                })?;
-
-            let keys: Vec<IdentityPublicKeyCreateTransition> = add_public_keys_array
-                .iter()
-                .map(|key| {
-                    let public_key: Ref<IdentityPublicKeyCreateTransitionWasm> =
-                        generic_of_js_val::<IdentityPublicKeyCreateTransitionWasm>(
-                            &key,
-                            "IdentityPublicKeyCreateTransition",
-                        )?;
-
-                    Ok(public_key.clone().into())
-                })
-                .collect::<Result<Vec<IdentityPublicKeyCreateTransition>, JsValue>>()?;
-
-            add_public_keys = Some(keys)
-        }
-
-        let mut disable_public_keys = None;
-
-        if js_sys::Reflect::has(public_keys, &"disable".into()).unwrap_or(false) {
-            let raw_disable_public_keys =
-                js_sys::Reflect::get(public_keys, &"disable".into()).unwrap();
-            let disable_public_keys_array: &js_sys::Array = raw_disable_public_keys
-                .dyn_ref::<js_sys::Array>()
-                .ok_or_else(|| {
-                    RustConversionError::Error(String::from("public keys to disable must be array"))
-                        .to_js_value()
-                })?;
-
-            let keys: Vec<KeyID> = disable_public_keys_array
-                .iter()
-                .map(|key| {
-                    let public_key_wasm: Ref<IdentityPublicKeyWasm> =
-                        generic_of_js_val::<IdentityPublicKeyWasm>(&key, "IdentityPublicKey")?;
-                    Ok(public_key_wasm.get_id())
-                })
-                .collect::<Result<Vec<KeyID>, JsValue>>()?;
-
-            disable_public_keys = Some(keys)
-        }
-
+        let (add_public_keys, disable_public_keys) =
+            super::factory_utils::parse_create_identity_update_transition_keys(public_keys)?;
         let now = js_sys::Date::now() as u64;
 
         self.0
