+++ packages/wasm-dpp/src/identity/mod.rs
@@ -1,3 +1,4 @@
+mod factory_utils;
 pub mod identity_facade;
 mod identity_factory;
 mod identity_public_key;
